#38 opts fixed

+++ pull.sh
@@ -21,8 +21,7 @@
 # OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
 # SOFTWARE.
 
-set +x
-set +e
+set -e
 
 # This script helps you create a pull request for Objectionary.
 # When you have a directory with sources, which you are ready
@@ -31,6 +30,14 @@ set +e
 # organized in "objectionary/eo-files" repository.
 
 repo=$1
+if [ "${repo}" == "" ]; then
+	echo "One argument is required as a name of GitHub repository"
+	echo "Read more here: https://github.com/objectionary/home"
+	exit -1
+fi
+
+set -x
+
 tmp=$(mktemp -d)
 trap 'rm -rf -- "${tmp}"' EXIT
